requirements.txt
scapy==2.5.0
requests==2.31.0
python-nmap==0.7.1
colorama==0.4.6